Check for a disconnected server; belay accessing data in various handlers.
This directly fixes #489.

This does not address fundamental flaws for destroying a resource in the
middle of a stacked sequence. Adding a destructive signal handler with
"signal_add_last()" only executes it last *at that stack level,* allowing
the previous level to continue ignorant of the destruction.

This "directly" fixes #489 because there are numerous other vectors for
destroying a resource (even this same one) in the middle of an event
sequence and no robust way to abort.

In the case of a single signal level: overuse of signal_stop() may prevent
future handlers from executing critical functions causing leaks and
regressions. In the case of multiple levels: signal_stop_by_name() perhaps
in server_disconnect(), in addition to the aforementioned, has no way to know
exactly what to stop and why, and how far up the stack to recurse to stop it.